refactor: BaseDialog 컴포넌트 core 모듈로 이동 및 공통 컴포넌트화 (#134)
- 모듈 재배치: `BaseDialog`를 `feature` 모듈에서 `core:designsystem` 모듈의 `component` 패키지로 이동
- 의존성 업데이트: `core:designsystem`에 Compose UI Tooling(Preview 및 Debug) 의존성 추가
- 참조 수정: `feature` 내 다이얼로그들(`UnsavedChangesDialog`, `ObjectDeleteConfirmDialog`, `SetCardTitleDialog`, `CardInfoDialog`)의 패키지 import 경로 변경
