+++ feature/src/main/java/com/zcard/feature/cardeditor/ui/dialog/CardInfoDialog.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.zcard.feature.R
-import com.zcard.feature.cardeditor.ui.common.BaseDialog
+import com.zcard.designsystem.component.BaseDialog
 import com.zcard.designsystem.theme.Gray
 import com.zcard.designsystem.theme.White
