+++ feature/src/main/java/com/zcard/feature/cardeditor/ui/dialog/ObjectDeleteConfirmDialog.kt
@@ -2,7 +2,7 @@ package com.zcard.feature.cardeditor.ui.dialog
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
-import com.zcard.feature.cardeditor.ui.common.BaseDialog
+import com.zcard.designsystem.component.BaseDialog
 import com.zcard.feature.R
 
 @Composable
